% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-debugging.R
\name{assert_input_type}
\alias{assert_input_type}
\title{assert_input_type}
\usage{
assert_input_type(object, types = character(0), tag = "")
}
\arguments{
\item{object}{some type of object}

\item{types}{list of object types acceptable}

\item{tag}{var to reference in error message}
}
\value{
no return value, run for side effects
}
\description{
assert_input_type
}
