% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Challenge2Data}
\alias{Challenge2Data}
\title{Plotting data Challenge 2}
\format{
An object of class \code{data.frame} with 113 rows and 5 columns.
}
\usage{
Challenge2Data
}
\description{
Plotting data Challenge 2
}
\keyword{datasets}
