% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cnvrt.coords}
\alias{cnvrt.coords}
\title{cnvrt.coords}
\usage{
cnvrt.coords(x, y = NULL, input = c("usr", "plt", "fig", "dev", "tdev"))
}
\arguments{
\item{x}{abscissa}

\item{y}{ordnate}

\item{input}{device}
}
\value{

}
\description{
Pulled from the TeachingDemos package (GPLv2+ Licensed)
}
\details{
Slightly hacked/trimmed
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>, Greg Snow <538280@gmail.com>
}
