% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-utilities.R
\name{cnvrt_coords}
\alias{cnvrt_coords}
\title{cnvrt_coords}
\usage{
cnvrt_coords(x, y = NULL, input = c("usr", "plt", "fig", "dev", "tdev"))
}
\arguments{
\item{x}{abscissa}

\item{y}{ordnate}

\item{input}{device}
}
\value{
Transformation of coordinates from local plot to figure space for phase changes
}
\description{
Pulled from the TeachingDemos package (GPLv2+ Licensed)
}
\details{
Slightly hacked/trimmed
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>, Greg Snow <538280@gmail.com>
}
