% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_brackets}
\alias{draw_brackets}
\title{draw_brackets}
\usage{
draw_brackets(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}
}
\value{

}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
