% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_cumsum_lines}
\alias{draw_cumsum_lines}
\title{draw_cumsum}
\usage{
draw_cumsum_lines(coreFrame, currentLayer, facetName)
}
\arguments{
\item{coreFrame}{fxl object}

\item{currentLayer}{layer to be drawn}

\item{facetName}{name of facet}
}
\value{

}
\description{
draw_cumsum
}
