% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_guide_line}
\alias{draw_guide_line}
\title{draw_guide_line}
\usage{
draw_guide_line(coreFrame, currentLayer, currentFacet)
}
\arguments{
\item{coreFrame}{fxl object}

\item{currentLayer}{layer to be drawn}

\item{n}{name of facet}
}
\value{

}
\description{
draw_guide_line
}
