% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-label-phase-draw.R
\name{draw_label_phase}
\alias{draw_label_phase}
\title{draw_label_phase}
\usage{
draw_label_phase(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}
}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
