% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-legend-draw.R
\name{draw_legend}
\alias{draw_legend}
\title{draw_legend}
\usage{
draw_legend(core_frame)
}
\arguments{
\item{core_frame}{fxl object}
}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
