% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-rect-draw.R
\name{draw_rect}
\alias{draw_rect}
\title{draw_rect}
\usage{
draw_rect(core_frame, current_layer, facet_name, zero_axis = FALSE)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}

\item{zero_axis}{filter out all but zeros}
}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
