% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-criterion-draw.R
\name{draw_scr_criterion}
\alias{draw_scr_criterion}
\title{draw_scr_criterion}
\usage{
draw_scr_criterion(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}
}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
