% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getFAMarkerColor}
\alias{getFAMarkerColor}
\title{getFAMarkerColor}
\usage{
getFAMarkerColor(condition)
}
\arguments{
\item{condition}{character(1) condition name}
}
\value{

}
\description{
Behavior analysts traditionally hold to conventions regarding specific
marker colors for FA conditions. This function attempts to match the
included conditions to a pre-specified colors.
}
