% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gilroyetal2019}
\alias{Gilroyetal2019}
\title{Plotting data from Gilroy et al. (2019a)}
\format{
An object of class \code{data.frame} with 15 rows and 9 columns.
}
\usage{
Gilroyetal2019
}
\description{
Plotting data from Gilroy et al. (2019a)
}
\keyword{datasets}
