% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gilroyetal2019Tx}
\alias{Gilroyetal2019Tx}
\title{Plotting data from Gilroy et al. (2019b)}
\format{
An object of class \code{data.frame} with 86 rows and 8 columns.
}
\usage{
Gilroyetal2019Tx
}
\description{
Plotting data from Gilroy et al. (2019b)
}
\keyword{datasets}
