% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gilroyetal2021}
\alias{Gilroyetal2021}
\title{Plotting data from Gilroy et al. (2021)}
\format{
An object of class \code{data.frame} with 69 rows and 7 columns.
}
\usage{
Gilroyetal2021
}
\description{
Plotting data from Gilroy et al. (2021)
}
\keyword{datasets}
