% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-debugging.R
\name{isValidAestheticMapping}
\alias{isValidAestheticMapping}
\title{isValidAestheticMapping}
\usage{
isValidAestheticMapping(object = NULL, name = NULL)
}
\arguments{
\item{object}{dataframe (hopefully)}

\item{name}{name for object}
}
\description{
isValidAestheticMapping
}
