% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-debugging.R
\name{isValidAXSCharacter}
\alias{isValidAXSCharacter}
\title{isValidAXSCharacter}
\usage{
isValidAXSCharacter(object = NULL, name = NULL)
}
\arguments{
\item{object}{some type of object}

\item{name}{parameter name}
}
\description{
isValidAXSCharacter
}
