% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-debugging.R
\name{isValidCharacterVector}
\alias{isValidCharacterVector}
\title{isValidCharacterVector}
\usage{
isValidCharacterVector(object = NULL, length = -1, name = NULL)
}
\arguments{
\item{object}{some type of object}

\item{length}{expected length}

\item{name}{parameter name}
}
\description{
isValidCharacterVector
}
