% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-debugging.R
\name{isValidDataFrame}
\alias{isValidDataFrame}
\title{isValidDataFrame}
\usage{
isValidDataFrame(object = NULL, name = NULL)
}
\arguments{
\item{object}{dataframe (hopefully)}

\item{name}{name for object}
}
\description{
isValidDataFrame
}
