% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-debugging.R
\name{isValidLogicalVector}
\alias{isValidLogicalVector}
\title{isValidLogicalVector}
\usage{
isValidLogicalVector(object = NULL, length = -1, name = NULL)
}
\arguments{
\item{object}{some type of object}

\item{length}{expected length}

\item{name}{parameter name}
}
\value{
no return value, run for side effects
}
\description{
isValidLogicalVector
}
