% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LozyEtAl2020}
\alias{LozyEtAl2020}
\title{Plotting data from Lozy et al. (2020)}
\format{
An object of class \code{data.frame} with 91 rows and 5 columns.
}
\usage{
LozyEtAl2020
}
\description{
Plotting data from Lozy et al. (2020)
}
\keyword{datasets}
