% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAnnotatedReversal}
\alias{plotAnnotatedReversal}
\title{plotMultiElementFA}
\usage{
plotAnnotatedReversal(
  data = NULL,
  grouping = NULL,
  session = NULL,
  response = NULL,
  response2 = NULL,
  response3 = NULL,
  condCol = NULL,
  pnum = NULL,
  poff = NULL,
  clabs = NULL,
  plabs = NULL,
  alines,
  abracks,
  title,
  ylab,
  xlab,
  ymins,
  ymaxs,
  deltaX
)
}
\arguments{
\item{data}{data.frame of assessment information}

\item{grouping}{character(1) column name, for grouping, referring to FA condition}

\item{session}{character(1) column name, for session numbering (should be distinct)}

\item{response}{character(1) column name, for response to chart}

\item{response2}{character(1) column name, for secondary response to chart}

\item{response3}{character(1) column name, for tertiary response to chart}

\item{condCol}{character(1) column name, for conditions}

\item{pnum}{character(1) column name, for numbered phases}

\item{poff}{character(1) column name, for offsetting overlapped phase connections}

\item{clabs}{data.frame condition labels}

\item{plabs}{data.frame phase labels}

\item{alines}{data.frame annotations for lines}

\item{abracks}{data.frame annotations for brackets}

\item{title}{character(1) plot title}

\item{ylab}{character(1) plot axis title}

\item{xlab}{character(1) plot axis title}

\item{ymins}{numeric(n) override ylims}

\item{ymaxs}{numeric(n) override ylims}

\item{deltaX}{numeric(n) n-by increases from 0 in sessions}
}
\value{

}
\description{
plotAnnotatedReversal
}
\details{
The goal of the plotAnnotatedReversal function is to construct a feature-
heavy figure that communicates many ecological changes without a table.
}
\examples{


# Specify location, text, and aesthetics of condition labels
conditionLabel <- data.frame(
  Panel = rep("John", 6),         # Only on top row (for andrew)
  X     = c(2.5,                  # Positioned between phase change lines
            6,
            9,
            12,
            15,
            18),
  Y     = rep(19, 6),             # Positioned near top of (John's) panel
  Cex   = rep(1.25, 6),           # Slightly oversized for clarity
  Text  = c("Baseline",           # Content of phase labels
            "FR-Lowest",
            "Baseline",
            "FR-Inelastic",
            "FR-Elastic",
            "FR-Inelastic")
)

# Specify location, text, and aesthetics of panel labels
panelLabel <- data.frame(
  Panel = c("John",               # Specify the panels to draw upon
            "Anthony",
            "Charles"),
  X     = rep(26, 3),             # Positioned near right of panel edge
  Y     = c(5,                    # Positioned at various heights, for each case
            12,
            21),
  Cex   = rep(1.5, 3),            # Slightly oversized for clarity
  Text  = c("John",               # Specific text to write
            "Anthony",
            "Charles")
)

# Specify if/how we want to show a legend
legendParams <- data.frame(
  panel =      "John",                # legend drawn on panel for John
  position =   "topright",            # drawn in top right corner
  legend = c("Responses Observed",    # Items to draw for legend
             "Reinforcers Produced"),
  col = c('black',                    # colors to draw lines
          'black'),
  lty = c(1, 2),                      # draw solid lines
  pch = c(19,2),                      # markers to draw
  bty = "n",                          # don't draw surrounding box
  pt.cex = 2.25,                      # oversize markers
  cex = 1.25,                         # slightly oversize text
  text.col = "black",                 # draw text in black
  horiz = F,                          # organize items vertically
  box.lty = 0                         # don't draw surrounding box
)

plotConcurrentReversals(data     = Gilroyetal2021,
                        grouping = 'Participant',
                        session  = 'Session',
                        response = 'Responding',
                        response2= 'Reinforcers',
                        pnum     = 'PhaseNum',
                        poff     = 'LineOff',
                        title    = 'Individual Evaluations of Reinforcer Efficacy and Elasticity across Reinforcers',
                        xlab     = 'Session',
                        ylab     = 'Frequency (Responses, Reinforcers Delivered)',
                        ymins    = list("John"    = 0,
                                        "Anthony" = 0,
                                        "Charles" = 0),
                        ymaxs    = list("John"    = 20,
                                        "Anthony" = 10,
                                        "Charles" = 20),
                        clabs    = conditionLabel,
                        plabs    = panelLabel,
                        llabs    = legendParams)
}
