% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotConcurrentReversals}
\alias{plotConcurrentReversals}
\title{plotConcurrentReversals}
\usage{
plotConcurrentReversals(
  data = NULL,
  grouping = NULL,
  session = NULL,
  response = NULL,
  response2 = NULL,
  pnum = NULL,
  poff = NULL,
  clabs = NULL,
  plabs = NULL,
  llabs = NULL,
  title,
  ylab,
  xlab,
  ymins,
  ymaxs
)
}
\arguments{
\item{data}{dataframe of assessment information}

\item{grouping}{character(1) column name, for grouping, referring to FA condition}

\item{session}{character(1) column name, for session numbering (should be distinct)}

\item{response}{character(1) column name, for response to chart}

\item{response2}{character(1) column name, for response to chart}

\item{pnum}{character(1) column name, for numbered phases}

\item{poff}{character(1) column name, for offsetting overlapped phase connections}

\item{clabs}{data.frame condition labels}

\item{plabs}{data.frame phase labels}

\item{llabs}{data.frame legend stuff}

\item{title}{character(1) plot title}

\item{ylab}{character(1) plot axis title}

\item{xlab}{character(1) plot axis title}

\item{ymins}{numeric(n) override ylims}

\item{ymaxs}{numeric(n) override ylims}
}
\value{

}
\description{
This is another multiple panel figure, however, conditions changes are not
always staggered out in terms of varying baselines. For example, this
implementation is more akin to a multiple probe design.
}
\examples{

conditionLabel <- data.frame(
Panel = rep("John", 6),
X     = c(2.5,
          6,
          9,
          12,
          15,
          18),
Y     = rep(19, 6),
Cex   = rep(1.25, 6),
Text  = c("Baseline",
          "FR-Lowest",
          "Baseline",
          "FR-Inelastic",
          "FR-Elastic",
          "FR-Inelastic")
)

panelLabel <- data.frame(
  Panel = c("John",
            "Anthony",
            "Charles"),
  X     = rep(26, 3),
  Y     = c(5,
            12,
            21),
  Cex   = rep(1.5, 3),
  Text  = c("John",
            "Anthony",
            "Charles")
)

legendParams <- data.frame(
  panel = "John",
  position = "topright",
  legend = c("Responses Observed", "Reinforcers Produced"),
  col = c('black', 'black'),
  lty = c(1, 2),
  pch = c(19,2),
  bty = "n",
  pt.cex = 2.25,
  cex = 1.25,
  text.col = "black",
  horiz = F,
  box.lty = 0
)

plotConcurrentReversals(data     = Gilroyetal2021,
                        grouping = 'Participant',
                        session  = 'Session',
                        response = 'Responding',
                        response2= 'Reinforcers',
                        pnum     = 'PhaseNum',
                        poff     = 'LineOff',
                        title    = 'Individual Evaluations of Reinforcer Efficacy and Elasticity across Reinforcers',
                        xlab     = 'Session',
                        ylab     = 'Frequency (Responses, Reinforcers Delivered)',
                        ymins    = list("John"    = 0,
                                        "Anthony" = 0,
                                        "Charles" = 0),
                        ymaxs    = list("John"    = 20,
                                        "Anthony" = 10,
                                        "Charles" = 20),
                        clabs    = conditionLabel,
                        plabs    = panelLabel,
                        llabs    = legendParams)

}
