% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMultiElementFA}
\alias{plotMultiElementFA}
\title{plotMultiElementFA}
\usage{
plotMultiElementFA(
  data = NULL,
  grouping = NULL,
  session = NULL,
  response = NULL,
  title,
  ylab,
  xlab,
  ylims,
  legend.horizonal = NULL,
  legend.position = NULL
)
}
\arguments{
\item{data}{data.frame of assessment information}

\item{grouping}{character(1) column name, for grouping, referring to FA condition}

\item{session}{character(1) column name, for session numbering (should be distinct)}

\item{response}{character(1) column name, for response to chart}

\item{title}{character(1) plot title}

\item{ylab}{character(1) plot axis title}

\item{xlab}{character(1) plot axis title}

\item{ylims}{numeric(n) override ylims}

\item{legend.horizonal}{logical(1) override legend aes}

\item{legend.position}{character(1) override legend aes}
}
\value{

}
\description{
plotMultiElementFA
}
\examples{

plotMultiElementFA(Gilroyetal2019,
grouping = 'Condition',
session  = 'Session',
response = 'CTB',
title    = 'Functional Analysis',
xlab     = 'Session',
ylab     = 'Frequency (Responses, Reinforcers Delivered)',
ylims    = c(0, 4),
legend.position = 'topright')

}
