% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMultipleBaseline}
\alias{plotMultipleBaseline}
\title{plotMultipleBaseline}
\usage{
plotMultipleBaseline(
  data = NULL,
  grouping = NULL,
  session = NULL,
  response = NULL,
  response2 = NULL,
  pnum = NULL,
  poff = NULL,
  clabs = NULL,
  plabs = NULL,
  llabs = NULL,
  title,
  ylab,
  xlab,
  ymins,
  ymaxs,
  xmax
)
}
\arguments{
\item{data}{dataframe of assessment information}

\item{grouping}{character(1) column name, for grouping, referring to FA condition}

\item{session}{character(1) column name, for session numbering (should be distinct)}

\item{response}{character(1) column name, for response to chart}

\item{response2}{character(1) column name, for response to chart}

\item{pnum}{character(1) column name, for numbered phases}

\item{poff}{character(1) column name, for offsetting overlapped phase connections}

\item{clabs}{data.frame condition labels}

\item{plabs}{data.frame phase labels}

\item{llabs}{data.frame legend stuff}

\item{title}{character(1) plot title}

\item{ylab}{character(1) plot axis title}

\item{xlab}{character(1) plot axis title}

\item{ymins}{numeric(n) override ylims}

\item{ymaxs}{numeric(n) override ylims}

\item{xmax}{numeric(1) override xlims}
}
\value{

}
\description{
plotMultipleBaseline
}
\examples{

conditionLabel <- data.frame(
Panel = rep("Andrew", 4),
X     = c(3.5,
          9,
          19,
          27),
Y     = rep(100, 4),
Cex   = rep(1.25, 4),
Text  = c("Baseline",
          "Training",
          "Post-Training",
          "Generalization")
)

panelLabel <- data.frame(
  Panel = c("Andrew",
            "Brian",
            "Charles"),
  X     = rep(27, 3),
  Y     = c(0,
            0,
            0),
  Cex   = rep(1.5, 3),
  Text  = c("Andrew",
            "Brian",
            "Charles")
)

# Roundabout way to 'prompt' drawing of lines in correct locations
dataFrameAdds <- data.frame(
  Participant	= c("Andrew",
                  "Brian",
                  "Charles"),
  Session	    = c(23,
                   23,
                   23),
  Condition	  = c("Generalization",
                  "Maintenance",
                  "Maintenance"),
  Responding	= c(NA,
                 NA,
                 NA),
  PhaseNum	  = c(3,3,4),
  LineOff     = c(0,0,0)
)

Gilroyetal2015 <- rbind(Gilroyetal2015,
                        dataFrameAdds)

plotMultipleBaseline(data     = Gilroyetal2015,
                     grouping = 'Participant',
                     session  = 'Session',
                     response = 'Responding',
                     pnum     = 'PhaseNum',
                     poff     = 'LineOff',
                     ymins    = list("Andrew"    = 0,
                                     "Brian" = 0,
                                     "Charles" = 0),
                     ymaxs    = list("Andrew"    = 100,
                                     "Brian" = 100,
                                     "Charles" = 100),
                     title    = 'Rates of Acquisition across Participants',
                     xlab     = 'Session',
                     xmax     = 27,
                     ylab     = 'Percent Accuracy',
                     clabs    = conditionLabel,
                     plabs    = panelLabel)

}
