% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-print.R
\name{print.fxl}
\alias{print.fxl}
\title{print.fxl}
\usage{
\method{print}{fxl}(x, ...)
}
\arguments{
\item{x}{fxl object}

\item{...}{inherits from generic}
}
\description{
Override the final call to print the fxl object. catches the obj and
prints out layers in the sequence laid out by the user
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
