% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-print-semilog.R
\name{print.fxlsemilog}
\alias{print.fxlsemilog}
\title{print.fxlsemilog}
\usage{
\method{print}{fxlsemilog}(x, ...)
}
\arguments{
\item{x}{fxlsemilog object}

\item{...}{inherits from generic}
}
\value{
no return, executed for side effects
}
\description{
Override the final call to print the fxl object.
catches the obj and prints out layers in the sequence laid out by the user
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
