% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-arrows.R
\name{scr_anno_arrows}
\alias{scr_anno_arrows}
\title{scr_anno_arrows}
\usage{
scr_anno_arrows(
  core_frame,
  arrows = NULL,
  facet = NULL,
  color = "black",
  length = 0.25,
  angle = 30,
  code = 2,
  lwd = 1,
  lty = 1
)
}
\arguments{
\item{core_frame}{fxl class}

\item{arrows}{list of keyed entries to be drawn on respective facets}

\item{facet}{the facet which will be drawn upon}

\item{color}{from base}

\item{length}{from base}

\item{angle}{from base}

\item{code}{from base}

\item{lwd}{from base}

\item{lty}{from base}
}
\description{
Add a layer with arrows to direct attention on the plot
}
\details{
Generally useful for avoiding a legend
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
