% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-guide-line.R
\name{scr_anno_guide_line}
\alias{scr_anno_guide_line}
\title{scr_anno_guide_line}
\usage{
scr_anno_guide_line(
  core_frame,
  coords,
  facet = NA,
  color = "black",
  lty = 1,
  lwd = 1
)
}
\arguments{
\item{core_frame}{fxl object}

\item{coords}{start and finish coords for aim line}

\item{facet}{panel to draw upon}

\item{color}{from base}

\item{lty}{line type}

\item{lwd}{line width}
}
\description{
This is an annotation illustrating an aim/reduction line
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
