% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-rect.R
\name{scr_anno_rect}
\alias{scr_anno_rect}
\title{scr_anno_rect}
\usage{
scr_anno_rect(core_frame, rects = NULL, color = "black", fill = "black")
}
\arguments{
\item{core_frame}{fxl object}

\item{rects}{list of keyed entries to be drawn on respective facets}

\item{color}{from base}

\item{fill}{from base}
}
\description{
scr_anno_rect
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
