% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-brackets.R
\name{scr_brackets}
\alias{scr_brackets}
\title{scr_brackets}
\usage{
scr_brackets(
  core_frame,
  brackets = NULL,
  facet = NULL,
  color = "black",
  length = 0.25,
  angle = 30,
  code = 2,
  lwd = 1,
  lty = 1
)
}
\arguments{
\item{core_frame}{fxl class}

\item{brackets}{list of keyed entries to be drawn on respective facets}

\item{facet}{the facet which will be drawn upon}

\item{color}{from base}

\item{length}{from base}

\item{angle}{from base}

\item{code}{from base}

\item{lwd}{from base}

\item{lty}{from base}
}
\value{

}
\description{
Add a layer with brackets on plot
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
