% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-criterion-line.R
\name{scr_criterion_lines}
\alias{scr_criterion_lines}
\title{scr_criterion_lines}
\usage{
scr_criterion_lines(
  core_frame,
  lty = 1,
  color = "black",
  size = 1,
  lines = NULL
)
}
\arguments{
\item{core_frame}{fxl object}

\item{lty}{from base}

\item{color}{from base}

\item{size}{from base}

\item{lines}{lines to draw}
}
\description{
scr_criterion_lines
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
