% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_cumsum_lines}
\alias{scr_cumsum_lines}
\title{scr_cumsum}
\usage{
scr_cumsum_lines(coreFrame, lty = 1, color = "black", size = 1, mapping)
}
\arguments{
\item{coreFrame}{fxl object}

\item{lty}{from base}

\item{color}{from base}

\item{size}{from base}

\item{mapping}{from base}
}
\value{

}
\description{
Draw lines, but as a cumulative and rolling sum
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
