% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_cumsum_points}
\alias{scr_cumsum_points}
\title{scr_cumsum_points}
\usage{
scr_cumsum_points(
  coreFrame,
  pch = 21,
  color = "black",
  fill = "black",
  cex = 1,
  mapping
)
}
\arguments{
\item{coreFrame}{fxl object}

\item{pch}{from base}

\item{color}{from base}

\item{fill}{from base}

\item{cex}{from base}

\item{mapping}{(optional) if overriding draw (i.e., different response)}
}
\value{

}
\description{
scr_cumsum_points
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
