% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-image.R
\name{scr_images}
\alias{scr_images}
\title{scr_images}
\usage{
scr_images(core_frame, image, cex = 1, mapping)
}
\arguments{
\item{core_frame}{fxl object}

\item{image}{should be RGML image}

\item{cex}{from base}

\item{mapping}{(optional) if overriding draw (i.e., different response)}
}
\value{
a layer to the core plotting object
}
\description{
scr_images
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
