% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_label_facet}
\alias{scr_label_facet}
\title{Title}
\usage{
scr_label_facet(coreFrame, color = "black", cex = 1, adj = 0.5, labels = NULL)
}
\arguments{
\item{coreFrame}{fxl object}

\item{color}{from base}

\item{cex}{from base}

\item{adj}{from base}

\item{labels}{as stated}
}
\value{

}
\description{
Title
}
