% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_label_phase}
\alias{scr_label_phase}
\title{scr_label_phase}
\usage{
scr_label_phase(
  coreFrame,
  color = "black",
  cex = 1,
  adj = 0.5,
  facet = NULL,
  labels = NULL
)
}
\arguments{
\item{coreFrame}{fxl object}

\item{color}{from base}

\item{cex}{from base}

\item{adj}{from base}

\item{facet}{facet of interest}

\item{labels}{as stated}
}
\value{

}
\description{
labels to be drawn on plots (typically for phases/conditions, but not necessarily)
}
