% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-legend.R
\name{scr_legend}
\alias{scr_legend}
\title{scrlegend}
\usage{
scr_legend(
  core_frame,
  panel = NA,
  legend,
  bg = NULL,
  col = NULL,
  pt_bg = NULL,
  lty,
  pch,
  box_lty = 0,
  adj = c(0, 0.5),
  bty = "n",
  cex = 1,
  horiz = FALSE,
  position = "topright",
  pt_cex = 1,
  text_col = "black",
  border = "black"
)
}
\arguments{
\item{core_frame}{fxl object}

\item{panel}{facet to be drawn on}

\item{legend}{from base}

\item{bg}{from base}

\item{col}{from base}

\item{pt_bg}{color, for point}

\item{lty}{from base}

\item{pch}{from base}

\item{box_lty}{from base}

\item{adj}{alignment}

\item{bty}{from base}

\item{cex}{from base}

\item{horiz}{from base}

\item{position}{from base}

\item{pt_cex}{from base}

\item{text_col}{from base}

\item{border}{border status (from base)}
}
\value{
nothing, run for side effects
}
\description{
Information for drawing legend onto plots
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
