% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_lines}
\alias{scr_lines}
\title{scr_lines}
\usage{
scr_lines(coreFrame, lty = 1, color = "black", size = 1, mapping)
}
\arguments{
\item{coreFrame}{fxl object}

\item{lty}{from base}

\item{color}{from base}

\item{size}{from base}

\item{mapping}{from base}
}
\value{

}
\description{
scr_lines
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
