% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_plines}
\alias{scr_plines}
\title{Title}
\usage{
scr_plines(coreFrame, lines = NULL, lwd = 1, lty = 1, col = "black")
}
\arguments{
\item{coreFrame}{fxl object}

\item{lines}{phase lines to be drawn}

\item{lwd}{from base}

\item{lty}{from base}

\item{col}{from base}
}
\value{

}
\description{
Title
}
