% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-rect.R
\name{scr_rect}
\alias{scr_rect}
\title{scr_rect}
\usage{
scr_rect(
  core_frame,
  x1 = NA,
  x2 = NA,
  y1 = NA,
  y2 = NA,
  facet = NA,
  color = "black",
  fill = "black"
)
}
\arguments{
\item{core_frame}{fxl object}

\item{x1}{from base}

\item{x2}{from base}

\item{y1}{from base}

\item{y2}{from base}

\item{facet}{from base}

\item{color}{from base}

\item{fill}{from base}
}
\description{
scr_rect
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
