% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-label-xlabel.R
\name{scr_xlabel}
\alias{scr_xlabel}
\title{xlabel}
\usage{
scr_xlabel(core_frame, var, line = 0)
}
\arguments{
\item{core_frame}{fxl object}

\item{var}{string}
}
\value{

}
\description{
Override the x axis label
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
