% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-axis-x.R
\name{scr_xoverride}
\alias{scr_xoverride}
\title{xoverride}
\usage{
scr_xoverride(
  core_frame,
  var,
  xdelta = 1,
  xticks = NULL,
  xdraws = NULL,
  xrotation = NULL,
  xtickscex = 1,
  xlabeloffset = NULL,
  xtickslabs = NULL,
  xticksadj = 1
)
}
\arguments{
\item{core_frame}{fxl object}

\item{var}{string for title}

\item{xdelta}{skips between ticks (can override)}

\item{xticks}{specify ticks, vector or named list}

\item{xdraws}{which x axes to draw}

\item{xrotation}{degree to rotate positioned labels}

\item{xtickscex}{expansion factor for labels}

\item{xlabeloffset}{offset to push labels downward}

\item{xtickslabs}{labels for x axis}

\item{xticksadj}{alignment for custom labels}
}
\value{
nothing, run for side effects
}
\description{
Override the x axis limits
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
