% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-label-ylabel.R
\name{scr_ylabel}
\alias{scr_ylabel}
\title{ylabel}
\usage{
scr_ylabel(
  core_frame,
  var,
  color = "black",
  cex = 1,
  adj = 0.5,
  face = 1,
  line = 0
)
}
\arguments{
\item{core_frame}{fxl object}

\item{var}{string}

\item{color}{from base}

\item{cex}{from base}

\item{adj}{from base}

\item{face}{like 'font' from base}

\item{line}{line width}
}
\value{
nothing, run for side effects
}
\description{
Override the y axis label
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
