% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimulatedAcademicFluency}
\alias{SimulatedAcademicFluency}
\title{Plotting data for Hypothetical Academic MTSS}
\format{
An object of class \code{data.frame} with 168 rows and 7 columns.
}
\usage{
SimulatedAcademicFluency
}
\description{
Plotting data for Hypothetical Academic MTSS
}
\keyword{datasets}
