% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-map.R
\name{var_map}
\alias{var_map}
\title{var_map}
\usage{
var_map(...)
}
\arguments{
\item{...}{map expressed relationships out}
}
\value{
list of exprs to map variables to plotting methods
}
\description{
This helper function maps out relationships to be parsed later on
}
