% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Challenge1Data}
\alias{Challenge1Data}
\title{Plotting data Challenge 1}
\format{
An object of class \code{data.frame} with 226 rows and 11 columns.
}
\usage{
Challenge1Data
}
\description{
Plotting data Challenge 1
}
\keyword{datasets}
