% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Challenge2Data}
\alias{Challenge2Data}
\title{Twitter chart challenge data 2}
\format{
A data frame with 113 rows and 5 variables:
\describe{
  \item{Participant}{Participant name}
  \item{Session}{Session number}
  \item{Condition}{Condition name}
  \item{IWS}{Incorrect word sequences}
  \item{CWS}{Correct word sounds}
}
}
\usage{
Challenge2Data
}
\description{
Twitter chart challenge data 2
}
\keyword{datasets}
