% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Challenge4Data}
\alias{Challenge4Data}
\title{Plotting data Challenge 4}
\format{
An object of class \code{data.frame} with 189 rows and 11 columns.
}
\usage{
Challenge4Data
}
\description{
Plotting data Challenge 4
}
\keyword{datasets}
