% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cnvrt.coords}
\alias{cnvrt.coords}
\title{cnvrt.coords (Modified)}
\usage{
cnvrt.coords(x, y = NULL, input = c("usr", "plt", "dev"))
}
\arguments{
\item{x}{numeric(1) abscissa index}

\item{y}{numeric(1) ordinate index}

\item{input}{character(1) device}
}
\value{
list: coordinate equivalents across multiple devices
}
\description{
A hacked-up, pared-down, reconstitution of helper methods to draw
using both user and figure-level coordinates
}
\details{
Author: Greg Snow <538280 at gmail.com>
Licensed: Artistic 2.0
Published: 2020-04-07
}
