% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-arrows-draw.R
\name{draw_arrows}
\alias{draw_arrows}
\title{draw_arrows}
\usage{
draw_arrows(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{n}{name of facet}
}
\value{

}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
