% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_bar_support}
\alias{draw_bar_support}
\title{draw_bar_support}
\usage{
draw_bar_support(coreFrame, currentLayer, facetName)
}
\arguments{
\item{coreFrame}{fxl object}

\item{currentLayer}{layer to be drawn}

\item{facetName}{name of facet}
}
\value{

}
\description{
Draw bars, but on a secondary axis
}
