% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-bar-draw.R
\name{draw_bar_support}
\alias{draw_bar_support}
\title{draw_bar_support}
\usage{
draw_bar_support(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}
}
\value{

}
\description{
Draw bars, but on a secondary axis
}
