% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-line-cumsum-line-draw.R
\name{draw_cumsum_lines}
\alias{draw_cumsum_lines}
\title{draw_cumsum_lines}
\usage{
draw_cumsum_lines(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}
}
\description{
draw_cumsum_lines
}
