% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-point-cumsum-point-draw.R
\name{draw_cumsum_points}
\alias{draw_cumsum_points}
\title{draw_cumsum_points}
\usage{
draw_cumsum_points(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}
}
\value{
nothing, run for side effects
}
\description{
draw_cumsum_points
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
