% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-guide-line-draw.R
\name{draw_guide_line}
\alias{draw_guide_line}
\title{draw_guide_line}
\usage{
draw_guide_line(core_frame, current_layer, facet_name)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}
}
\value{

}
\description{
draw_guide_line
}
