% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_label_facet}
\alias{draw_label_facet}
\title{draw_label_facet}
\usage{
draw_label_facet(coreFrame, currentLayer, n)
}
\arguments{
\item{coreFrame}{fxl object}

\item{currentLayer}{layer to be drawn}

\item{n}{name of facet}
}
\value{

}
\description{
drawing function
}
