% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-point-draw.R
\name{draw_points}
\alias{draw_points}
\title{draw_points}
\usage{
draw_points(core_frame, current_layer, facet_name, zero_axis = FALSE)
}
\arguments{
\item{core_frame}{fxl object}

\item{current_layer}{layer to be drawn}

\item{facet_name}{name of facet}

\item{zero_axis}{filter out all but zeros}
}
\value{
nothing, run for side effects
}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
