% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{draw_scr_plines}
\alias{draw_scr_plines}
\title{draw_scr_plines}
\usage{
draw_scr_plines(coreFrame, currentLayer, n)
}
\arguments{
\item{coreFrame}{fxl object}

\item{currentLayer}{layer to be drawn}

\item{n}{name of facet}
}
\value{

}
\description{
drawing function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
