% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GelinoEtAl2022}
\alias{GelinoEtAl2022}
\title{Plotting data from Gelino et al. (2022)}
\format{
An object of class \code{data.frame} with 11 rows and 9 columns.
}
\usage{
GelinoEtAl2022
}
\description{
Plotting data from Gelino et al. (2022)
}
\keyword{datasets}
