% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gilroyetal2015}
\alias{Gilroyetal2015}
\title{Plotting data from Gilroy et al. (2015)}
\format{
An object of class \code{data.frame} with 40 rows and 6 columns.
}
\usage{
Gilroyetal2015
}
\description{
Plotting data from Gilroy et al. (2015)
}
\keyword{datasets}
