% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gilroyetal2015}
\alias{Gilroyetal2015}
\title{Plotting data from Gilroy et al. (2015)}
\format{
A data frame with 40 rows and 6 variables:
\describe{
  \item{Participant}{Participant name}
  \item{Session}{Session number}
  \item{Condition}{Condition name}
  \item{Responding}{Responding rates}
  \item{PhaseNum}{Phase number}
  \item{LineOff}{Offset for phase line}
}
}
\source{
<doi:https://doi.org/10.1016/j.rasd.2015.04.004>
}
\usage{
Gilroyetal2015
}
\description{
This is data either extracted or included (with permission) to assist with illustrating and replicating the functionality of the package. Specifically, this data is used to illustrate conventions regarding a multiple probe design and how that may be constructed in this package.
}
\keyword{datasets}
