% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gilroyetal2019}
\alias{Gilroyetal2019}
\title{Plotting data from Gilroy et al. (2019) - FA}
\format{
A data frame with 15 rows and 9 variables:
\describe{
  \item{Session}{Session number}
  \item{SIB}{Rates of self-injury}
  \item{AGG}{Rates of aggression}
  \item{DIS}{Rates of disruptive behavior}
  \item{Prompt}{Rates of prompting}
  \item{Comply}{Rates of compliance}
  \item{SR}{Duration of reinforcement}
  \item{CTB}{Rates of combined target behavior}
  \item{Condition}{Functional analysis condition}
}
}
\source{
<doi:https://doi.org/10.1080/17518423.2019.1646342>
}
\usage{
Gilroyetal2019
}
\description{
FA data from Gilroy et al. (2019)
}
\keyword{datasets}
