% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-debugging.R
\name{isValidNumericVector}
\alias{isValidNumericVector}
\title{isValidNumericVector}
\usage{
isValidNumericVector(object, length = -1, name = NULL)
}
\arguments{
\item{object}{some type of object}

\item{length}{expected length}

\item{name}{parameter name}
}
\description{
isValidNumericVector
}
