% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KoffarnusEtAl2011}
\alias{KoffarnusEtAl2011}
\title{Plotting data from Koffarnus et al. (2011)}
\format{
An object of class \code{data.frame} with 14979 rows and 3 columns.
}
\usage{
KoffarnusEtAl2011
}
\description{
Plotting data from Koffarnus et al. (2011)
}
\keyword{datasets}
