% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LozyEtAl2020}
\alias{LozyEtAl2020}
\title{Plotting data from Lozy et al. (2020)}
\format{
A data frame with 91 rows and 5 variables:
\describe{
  \item{Session}{Session number}
  \item{Participant}{Participant name}
  \item{KM}{Kinesthetic movement choices}
  \item{TD}{Traditional drill choices}
  \item{Phase}{Sequenced phase number}
}
}
\source{
<doi:https://doi.org/10.1002/jaba.677>
}
\usage{
LozyEtAl2020
}
\description{
Treatment data from Lozy et al. (2020)
}
\keyword{datasets}
