% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMultiElementFA}
\alias{plotMultiElementFA}
\title{plotMultiElementFA}
\usage{
plotMultiElementFA(
  data = NULL,
  grouping = NULL,
  session = NULL,
  response = NULL,
  title,
  ylab,
  xlab,
  ylims,
  legend.horizonal = NULL,
  legend.position = NULL
)
}
\arguments{
\item{data}{data frame of assessment information}

\item{grouping}{column name, for grouping, referring to FA condition}

\item{session}{column name, for session numbering (should be distinct)}

\item{response}{column name, for response to chart}

\item{title}{plot title}

\item{ylab}{plot axis title}

\item{xlab}{plot axis title}

\item{ylims}{override ylims}

\item{legend.horizonal}{override legend aes}

\item{legend.position}{override legend aes}
}
\value{

}
\description{
plotMultiElementFA
}
\examples{

plotMultiElementFA(Gilroyetal2019,
grouping = 'Condition',
session  = 'Session',
response = 'CTB',
title    = 'Functional Analysis',
xlab     = 'Session',
ylab     = 'Frequency (Responses, Reinforcers Delivered)',
ylims    = c(0, 4),
legend.position = 'topright')

}
