% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxl-print.R
\name{print.fxl}
\alias{print.fxl}
\title{print.fxl}
\usage{
\method{print}{fxl}(core_frame, ...)
}
\arguments{
\item{core_frame}{fxl object}
}
\value{

}
\description{
Override the final call to print the fxl object. catches the obj and
prints out layers in the sequence laid out by the user
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
