% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{print.fxlsemilog}
\alias{print.fxlsemilog}
\title{print.fxlsemilog}
\usage{
\method{print}{fxlsemilog}(core_frame, ...)
}
\arguments{
\item{core_frame}{fxlsemilog object}
}
\value{

}
\description{
Override the final call to print the fxl object. catches the obj and prints out layers in the sequence laid out by the user
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
