% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-bar.R
\name{scr_bar_support}
\alias{scr_bar_support}
\title{scr_bar_support}
\usage{
scr_bar_support(
  core_frame,
  color = rgb(0.8, 0.8, 0.8, alpha = 0.25),
  alpha = 1,
  mapping = NULL,
  label = "",
  width = 0.8
)
}
\arguments{
\item{core_frame}{fxl object}

\item{color}{from base}

\item{alpha}{from base}

\item{mapping}{(optional) if overriding draw (i.e., different response)}

\item{label}{description for bar}

\item{width}{width of bar}
}
\description{
Adds a supplemental bar to the figure, if relevant to the data
}
