% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-support-bar.R
\name{scr_bar_support}
\alias{scr_bar_support}
\title{scr_bar_support}
\usage{
scr_bar_support(
  core_frame,
  color = rgb(0.8, 0.8, 0.8, alpha = 0.25),
  alpha = 1,
  guide_line = NULL,
  guide_line_type = 1,
  guide_line_size = 1,
  guide_line_color = "black",
  mapping = NULL,
  label = "",
  styler = NA,
  width = 0.8
)
}
\arguments{
\item{core_frame}{fxl object}

\item{color}{from base}

\item{alpha}{from base}

\item{guide_line}{(optional) aim line for bars}

\item{guide_line_type}{(optional) aim line type for bars}

\item{guide_line_size}{(optional) aim line size for bars}

\item{guide_line_color}{(optional) aim line color for bars}

\item{mapping}{(optional) if overriding draw (i.e., different response)}

\item{label}{description for bar}

\item{styler}{a lambda function that returns dynamic styling parameters}

\item{width}{width of bar}
}
\description{
Adds a supplemental bar to the figure, if relevant to the data
}
