% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_bar_support}
\alias{scr_bar_support}
\title{scr_bar_support}
\usage{
scr_bar_support(
  coreFrame,
  color = rgb(0.8, 0.8, 0.8, alpha = 0.25),
  alpha = 1,
  mapping = NULL,
  label = "",
  width = 0.8
)
}
\arguments{
\item{coreFrame}{fxl object}

\item{color}{from base}

\item{alpha}{from base}

\item{mapping}{(optional) if overriding draw (i.e., different response)}
}
\value{

}
\description{
Adds a supplemental bar to the figure, if relevant to the data
}
