% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-label-facet.R
\name{scr_label_facet}
\alias{scr_label_facet}
\title{scr_label_facet}
\usage{
scr_label_facet(
  core_frame,
  color = "black",
  cex = 1,
  adj = 0.5,
  face = 1,
  x = NULL,
  y = NULL,
  labels = NULL
)
}
\arguments{
\item{core_frame}{fxl object}

\item{color}{from base}

\item{cex}{from base}

\item{adj}{from base}

\item{face}{like 'font' from base}

\item{x}{global x position for labels}

\item{y}{global y position for labels}

\item{labels}{as stated}
}
\description{
scr_label_facet
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
