% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-label-phase.R
\name{scr_label_phase}
\alias{scr_label_phase}
\title{scr_label_phase}
\usage{
scr_label_phase(
  core_frame,
  color = "black",
  cex = 1,
  adj = 0.5,
  face = 1,
  x = NULL,
  y = NULL,
  facet = NULL,
  labels = NULL
)
}
\arguments{
\item{core_frame}{fxl object}

\item{color}{from base}

\item{cex}{from base}

\item{adj}{from base}

\item{face}{like 'font' from base}

\item{x}{location}

\item{y}{location}

\item{facet}{facet of interest}

\item{labels}{as stated}
}
\description{
labels to be drawn on plots (typically for phases/conditions, but not necessarily)
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
