% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_legend}
\alias{scr_legend}
\title{scrlegend}
\usage{
scr_legend(
  coreFrame,
  panel = NA,
  legend,
  col,
  lty,
  pch,
  box.lty = 0,
  bty = "n",
  cex = 1,
  horiz = FALSE,
  position = "topright",
  pt.cex = 1,
  text.col = "black"
)
}
\arguments{
\item{coreFrame}{fxl object}

\item{panel}{facet to be drawn on}

\item{legend}{from base}

\item{col}{from base}

\item{lty}{from base}

\item{pch}{from base}

\item{box.lty}{from base}

\item{bty}{from base}

\item{cex}{from base}

\item{horiz}{from base}

\item{position}{from base}

\item{pt.cex}{from base}

\item{text.col}{from base}
}
\value{

}
\description{
Information for drawing legend onto plots
}
