% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-phasechange-single.R
\name{scr_plines}
\alias{scr_plines}
\title{scr_plines}
\usage{
scr_plines(core_frame, lines = NULL, lwd = 1, lty = 1, col = "black")
}
\arguments{
\item{core_frame}{fxl object}

\item{lines}{phase lines to be drawn}

\item{lwd}{from base}

\item{lty}{from base}

\item{col}{from base}
}
\value{
a layer to the core plotting object
}
\description{
scr_plines
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
