% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-phasechange-multiple.R
\name{scr_plines_mbd}
\alias{scr_plines_mbd}
\title{scr_plines_mbd}
\usage{
scr_plines_mbd(core_frame, lines = NULL)
}
\arguments{
\item{core_frame}{fxl object}

\item{lines}{phase lines to be drawn}
}
\description{
scr_plines_mbd
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
