% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_plines_mbd}
\alias{scr_plines_mbd}
\title{scr_plines_mbd}
\usage{
scr_plines_mbd(coreFrame, lines = NULL)
}
\arguments{
\item{coreFrame}{fxl object}

\item{lines}{phase lines to be drawn}
}
\value{

}
\description{
scr_plines_mbd
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
