% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_plot}
\alias{scr_plot}
\title{scrplot}
\usage{
scr_plot(
  data,
  aesthetics,
  mai = c(0.375, 0.375, 0.25, 0.25),
  omi = c(0.25, 0.25, 0.25, 0.25),
  ncol = 1,
  family = NULL,
  semilog = FALSE
)
}
\arguments{
\item{data}{submitted data (not opinionated on naming)}

\item{aesthetics}{references for data in frame}

\item{mai}{margins in inches}

\item{omi}{outer margins in inches}
}
\value{

}
\description{
Core object for establishing fxl object and layers
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
