% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-point.R
\name{scr_points}
\alias{scr_points}
\title{scr_points}
\usage{
scr_points(
  core_frame,
  pch = 21,
  color = "black",
  fill = "black",
  cex = 1,
  mapping
)
}
\arguments{
\item{core_frame}{fxl object}

\item{pch}{from base}

\item{color}{from base}

\item{fill}{from base}

\item{cex}{from base}

\item{mapping}{(optional) if overriding draw (i.e., different response)}
}
\description{
scr_points
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
