% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-anno-rect.R
\name{scr_rect}
\alias{scr_rect}
\title{scr_rect}
\usage{
scr_rect(core_frame, rects = NULL, facet = NA, color = "black", fill = "black")
}
\arguments{
\item{core_frame}{fxl object}

\item{rects}{list of keyed entries to be drawn on respective facets}

\item{facet}{the facet which will be drawn upon}

\item{color}{from base}

\item{fill}{from base}

\item{x1}{from base}

\item{x2}{from base}

\item{y1}{from base}

\item{y2}{from base}
}
\description{
scr_rect
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
