% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_save}
\alias{scr_save}
\title{scrsave}
\usage{
scr_save(
  coreFrame,
  units = "in",
  name = "test.tiff",
  width = 8,
  height = 4,
  res = 600
)
}
\arguments{
\item{coreFrame}{fxl object}

\item{units}{from base}

\item{name}{from base}

\item{width}{from base}

\item{height}{from base}

\item{res}{from base}
}
\value{

}
\description{
Function for outputting fxl object at preset size (certain journal are opinionated on size, format, and density)
}
