% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_title}
\alias{scr_title}
\title{scrtitle}
\usage{
scr_title(coreFrame, var)
}
\arguments{
\item{coreFrame}{fxl object}

\item{var}{string}
}
\value{

}
\description{
Override the title
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
