% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-label-xlabel.R
\name{scr_xlabel}
\alias{scr_xlabel}
\title{xlabel}
\usage{
scr_xlabel(
  core_frame,
  var,
  color = "black",
  cex = 1,
  adj = 0.5,
  face = 1,
  line = 0
)
}
\arguments{
\item{core_frame}{fxl object}

\item{var}{string}

\item{color}{from base}

\item{cex}{from base}

\item{adj}{from base}

\item{face}{like 'font' from base}

\item{line}{line width}
}
\value{
nothing, run for side effects
}
\description{
Override the x axis label
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
