% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_ylabel}
\alias{scr_ylabel}
\title{ylabel}
\usage{
scr_ylabel(core_frame, var, line = 0)
}
\arguments{
\item{core_frame}{fxl object}

\item{var}{string}
}
\value{

}
\description{
Override the y axis label
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
