% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-axis-y.R
\name{scr_yoverride}
\alias{scr_yoverride}
\title{yoverride}
\usage{
scr_yoverride(core_frame, var, ydelta = 1, ydraws = NULL)
}
\arguments{
\item{core_frame}{fxl object}

\item{var}{from base}

\item{ydelta}{skips between ticks (can override)}

\item{ydraws}{specify axes manual}
}
\value{

}
\description{
Override the y axis (or axes) limits
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
