% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fxl.R
\name{scr_yoverride}
\alias{scr_yoverride}
\title{yoverride}
\usage{
scr_yoverride(coreFrame, var, ydelta = 1)
}
\arguments{
\item{coreFrame}{fxl object}

\item{var}{from base}

\item{ydelta}{skips between ticks (can override)}
}
\value{

}
\description{
Override the y axis (or axes) limits
}
