% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimulatedAcademicFluency}
\alias{SimulatedAcademicFluency}
\title{Plotting data for Hypothetical Academic MTSS}
\format{
A data frame with 168 rows and 7 variables:
\describe{
  \item{Rates}{Rates of change}
  \item{Times}{Multiplier for model level}
  \item{index}{Individual id}
  \item{starts}{Modeled baseline start}
  \item{jitter}{Jitter offset}
  \item{pred}{Prediction from model}
  \item{err}{Residual error}
}
}
\usage{
SimulatedAcademicFluency
}
\description{
Plotting data for Hypothetical Academic MTSS
}
\keyword{datasets}
